API

In the following, you find the documentation of some, and hopefully soon all, exported functions of MNPDynamics.jl:

MNPDynamics.simulationMNP

MNPDynamics.simulationMNP

MNPDynamics.simulationMNPMultiParams

MNPDynamics.simulationMNPMultiParams

MNPDynamics.simulationMNP — Method

simulationMNP(B::Matrix{T}, t; kargs...)

This version takes the fields in discretized form

MNPDynamics.simulationMNP — Method

simulationMNP(B, t, offsets; kargs...)

MNPDynamics.simulationMNPMultiParams — Method

simulationMNPMultiParams(B, t, params; kargs...)

MNPDynamics.simulationMNPMultiParams — Method

simulationMNPMultiParams(B::Vector{Matrix{T}}, t; kargs...)

This version takes the fields in discretized form