API

In the following, you find the documentation of some, and hopefully soon all, exported functions of MNPDynamics.jl:

MNPDynamics.simulationMNP

MNPDynamics.simulationMNPMultiOffsets

MNPDynamics.simulationMNP — Method

simulationMNP(B, t, offsets; kargs...)

MNPDynamics.simulationMNPMultiOffsets — Method

simulationMNPMultiOffsets(B, t, offsets; kargs...)